Ultimos cambios, pasamos la clase :,c

+++ app/Http/Controllers/PedidosController.php
@@ -103,7 +103,10 @@ class PedidosController extends Controller
     public function destroy(int $pedidosId)
     {
         //
+        
         $pedidos=Pedidos::find($pedidosId);
+        dd($pedidos);
+        $this->authorize('delete', $pedidos);
         $pedidos->delete();
         return redirect('/pedidos');
     }
